[Fix] Position Simulation: frame is int, not float

+++ src/drones_manager/trajectory_simulation_manager/trajectory_simulation_manager.py
@@ -7,7 +7,7 @@ from ..drone.events.position_events import PositionEvent
 
 @dataclass(frozen=True)
 class PositionSimulation:
-    frame: float
+    frame: int
     xyz: Tuple[float, float, float]
